마이페이지 기능 완성 및 최근 본 게시글, 클래스 분할 및 게시글 속 닉네임 표기
