+++ lib/post/post_search.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:cloud_firestore/cloud_firestore.dart';
+import 'package:oneroom_finder/userinfo/recentlyviewed.dart';
 import 'room_details_screen.dart';
 
 class PostSearchDelegate extends SearchDelegate<String> {
